Refine SRE interview instructions
Reduce snags/ambiguity.

We're changing the port number since, on the ephemeral EC2s created as
part of the test, we were publicly exposing the SSH server even though
there was no functional reason to do so. Vulnerabilities like
regreSSHion tell us we shouldn't be exposing services unless we have a
good reason to do so. As such, the port has been changed to something
with no service behind it and that is firewalled by default. It also
amuses me.